Dropping shib login error message in DSpaceXmluiApi in favor of #539
