+++ dspace-xmlui/src/main/java/cz/cuni/mff/ufal/DSpaceXmluiApi.java
@@ -55,89 +55,6 @@ public class DSpaceXmluiApi {
 			div.addPara("errors", "alert alert-danger").addContent(err);
 		}
 	}
-	
-	public static void app_xmlui_aspect_eperson_postError(Context context, Message head, Division div)
-			throws org.dspace.app.xmlui.wing.WingException{
-
-		Object o = context.fromCache(cz.cuni.mff.ufal.Headers.class, 1);
-		if ( o != null ) 
-		{
-			final Map<String, java.util.List<String>> headers = ((cz.cuni.mff.ufal.Headers)o).get();
-			div.setHead("Required contact details missing");
-			String feedback = ConfigurationManager.getProperty("feedback.recipient");
-			if (org.apache.commons.lang.StringUtils.isBlank(feedback)){
-				log.error("feedback.recipient not set.");
-			}
-			Para para = div.addPara();
-			
-			String ourEntityId = ConfigurationManager.getProperty("authentication-shibboleth","spEntityId");
-			
-			String idp = find_idp(headers);
-			para.addContent(
-					String.format(
-					"The authentication was successful; however, your identity provider (in most cases " +
-					"it is your home institution - %s) did provide neither your " +
-					"email, eppn (specific authentication id) nor targeted id (specific authentication id). " +
-					"We need at least one of the properties mentioned above so our service can work " +
-					"properly.\nTry to contact your identity provder and ask them for these attributes. If they ask for our entityID it is \"" + 
-					ourEntityId + "\". " +
-					"\n\nFor more information please contact us at ", idp));
-			para.addXref("mailto:"+feedback+"?subject=My IDP does not provide an email", feedback);
-			para.addContent(". We might arrange a local account for you or try to resolve the issue with your identity provider. " +
-					"Please include the following details in the email:");
-			
-			para = div.addPara("show-attr", null);
-			para.addContent("Show details");
-			
-			Table table = div.addTable("shib-attr", 1, 1);			
-			for(String headerName : headers.keySet() ){
-				Row row = table.addRow();
-				row.addCellContent(headerName);
-				Cell cell = null;
-				if(headerName.matches(".*mail.*")){
-					cell = row.addCell(null, null, HtmlHelper.cls(HtmlHelper.header_class.NOT_OK));
-				}
-				else{
-					cell = row.addCell();
-				}
-				cell.addContent((headers.get(headerName)).toString().replaceFirst("\\[(.*)\\]", "$1").replaceAll(";", "; "));
-			}
-			
-			Division dialog = div.addDivision("dialog");
-			dialog.addPara("The authentication was successful but " +
-					"we have not received all the required information. " +
-					"Please, read carefully the information on the following page. " +
-					"If you need more help, do not hesitate to contact us using the provided link.");
-			/*List list = dialog.addList("dialog-list", org.dspace.app.xmlui.wing.element.List.TYPE_ORDERED);
-			list.addItem("Contact Your IDP");
-			list.addItem("If Your IDP can't or refuses to help you, contact us");*/
-			return;
-		}
-		
-		if(head != null){
-			div.setHead(head);
-		}			
-		app_xmlui_aspect_eperson_postError(div);
-
-	}
-	
-	private static String find_idp( Map<String, java.util.List<String>> headers ) {
-		java.util.List<String> ret = find_shibboleth_header_value(headers, "Shib-Identity-Provider");
-		return ret == null ? "unknown" : ret.get(0);
-	}
-	private static java.util.List<String> find_shibboleth_header_value( 
-			Map<String, java.util.List<String>> headers,
-			String to_find) 
-	{
-		to_find = to_find.toLowerCase();
-		for ( String key : headers.keySet() )
-		{
-			if ( key.toLowerCase().equals(to_find) ) {
-				return headers.get(key);
-			}
-		}
-		return null;
-	}
 
 	/*
 	 * Function for app.xmlui.aspect.artifactbrowser.ItemViewer.java. This
