+++ dspace-xmlui/src/main/webapp/themes/UFAL/lib/html/cs/error.html
@@ -43,6 +43,7 @@
                     "Please contact our helpdesk": "Kontaktujte prosím poradnu",
                     "Cannot continue with authentication": "V přihlášení nelze pokračovat",
                     "The login process can\'t continue because your home institution (%s) did not send the required information. Please click the button and send the prepared email to the responsible person from your IT department. You\'ll be also helping your colleagues.": "V přihlášení nelze pokračovat, protože vaše domovská instituce (%s) neposkytla potřebné informace. Klikněte prosím na tlačítko a odešlete připravený e-mail zodpovědné osobě vašeho IT oddělení. Pomůžete tím i vašim kolegům.",
+                    "The login process can\'t continue because your home institution (%s) did not send the required information. Please click the button and send us an email. You\'ll be also helping your colleagues.":"V přihlášení nelze pokračovat, protože vaše domovská instituce (%s) neposkytla potřebné informace. Klikněte prosím na tlačítko a pošlete nám e-mail. Pomůžete tím i vašim kolegům.",
                     "Send Email": "Poslat e-mail",
                     "To": "Na adresu",
                     "Subject": "Předmět",
